+++ src/main/java/com/filmdoms/community/account/data/dto/response/AccountResponseDto.java
@@ -1,7 +1,7 @@
 package com.filmdoms.community.account.data.dto.response;
 
-import com.filmdoms.community.account.data.constants.AccountRole;
-import com.filmdoms.community.account.data.constants.AccountStatus;
+import com.filmdoms.community.account.data.constant.AccountRole;
+import com.filmdoms.community.account.data.constant.AccountStatus;
 import com.filmdoms.community.account.data.entity.Account;
 import com.filmdoms.community.file.data.dto.response.FileResponseDto;
 import lombok.AccessLevel;
