+++ src/main/java/com/filmdoms/community/account/service/AccountService.java
@@ -1,7 +1,7 @@
 package com.filmdoms.community.account.service;
 
 import com.filmdoms.community.account.config.jwt.JwtTokenProvider;
-import com.filmdoms.community.account.data.constants.AccountRole;
+import com.filmdoms.community.account.data.constant.AccountRole;
 import com.filmdoms.community.account.data.dto.AccountDto;
 import com.filmdoms.community.account.data.dto.request.DeleteAccountRequestDto;
 import com.filmdoms.community.account.data.dto.request.JoinRequestDto;
