+++ src/main/java/com/filmdoms/community/article/service/ArticleService.java
@@ -27,8 +27,8 @@ import com.filmdoms.community.article.repository.CriticRepository;
 import com.filmdoms.community.article.repository.FilmUniverseRepository;
 import com.filmdoms.community.file.data.entity.File;
 import com.filmdoms.community.file.repository.FileRepository;
-import com.filmdoms.community.imagefile.service.ImageFileService;
-import com.filmdoms.community.newcomment.repository.NewCommentRepository;
+import com.filmdoms.community.file.service.ImageFileService;
+import com.filmdoms.community.comment.repository.CommentRepository;
 import com.filmdoms.community.vote.data.entity.VoteKey;
 import com.filmdoms.community.vote.repository.VoteRepository;
 import lombok.RequiredArgsConstructor;
@@ -51,7 +51,7 @@ public class ArticleService {
     private final FilmUniverseRepository filmUniverseRepository;
     private final CriticRepository criticRepository;
     private final FileRepository fileRepository;
-    private final NewCommentRepository newCommentRepository;
+    private final CommentRepository commentRepository;
     private final ImageFileService imageFileService;
     private final VoteRepository voteRepository;
     private final AccountRepository accountRepository;
