+++ src/main/java/com/filmdoms/community/board/data/BoardHeadCore.java
@@ -1,23 +1,9 @@
 package com.filmdoms.community.board.data;
 
 import com.filmdoms.community.account.data.entity.Account;
-import com.filmdoms.community.board.data.constant.PostStatus;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.DiscriminatorColumn;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Inheritance;
-import jakarta.persistence.InheritanceType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
+import com.filmdoms.community.account.data.entity.BaseTimeEntity;
+import com.filmdoms.community.article.data.constant.PostStatus;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
