+++ src/main/java/com/filmdoms/community/board/post/service/PostCommentService.java
@@ -1,7 +1,5 @@
 package com.filmdoms.community.board.post.service;
 
-import com.filmdoms.community.account.data.constants.AccountRole;
-import com.filmdoms.community.account.data.entity.Account;
 import com.filmdoms.community.account.repository.AccountRepository;
 import com.filmdoms.community.board.post.data.entity.PostComment;
 import com.filmdoms.community.board.post.repository.PostCommentRepository;
